+++ src/components/LiquidityDepth/DensityChart.tsx
@@ -105,8 +105,8 @@ export default function DensityChart({
           ) : (
             <LiquidityChartRangeInput
               data={{ series: filteredData, current: parseFloat(price) }}
-              dimensions={{ width: 350, height: 200 }}
-              margins={{ top: 0, right: 0, bottom: 0, left: 0 }}
+              dimensions={{ width: 350, height: 250 }}
+              margins={{ top: 20, right: 20, bottom: 20, left: 20 }}
               styles={{
                 area: {
                   fill: theme.blue1,
