+++ src/pages/AddLiquidity/index.tsx
@@ -654,9 +654,6 @@ export default function AddLiquidity({
                       priceLabel={
                         price && baseCurrency && quoteCurrency && !hasExistingPosition && !noLiquidity ? (
                           <AutoRow gap="4px" justify="center">
-                            <TYPE.body color="text2" fontSize={12}>
-                              <Trans>Current Price:</Trans>
-                            </TYPE.body>
                             <TYPE.body fontWeight={500} color="text2" fontSize={12}>
                               <HoverInlineText
                                 maxCharacters={20}
